Refactor publish script
